+++ gsplat/cuda/_wrapper.py
@@ -2,18 +2,6 @@ from typing import Callable, Optional, Tuple
 
 import torch
 from torch import Tensor
-import math
-
-N_max = 51
-binoms = torch.zeros((N_max, N_max)).float().cuda()
-for n in range(N_max):
-    for k in range(n+1):
-        binoms[n, k] = math.comb(n, k)
-
-def compute_relocation_cuda(opacity_old, scale_old, N):
-    from ._backend import _C
-    new_opacity, new_scale = _C.compute_relocation(opacity_old, scale_old, N.int(), binoms, N_max)
-    return new_opacity, new_scale 
 
 def _make_lazy_cuda_func(name: str) -> Callable:
     def call_cuda(*args, **kwargs):
@@ -23,7 +11,6 @@ def _make_lazy_cuda_func(name: str) -> Callable:
         return getattr(_C, name)(*args, **kwargs)
 
     return call_cuda
-
 
 def spherical_harmonics(
     degrees_to_use: int,
